add crud for admin

+++ views/departement.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, jsonify, request
-from models import Departement, DepartementSchema, db
+from models import Departement, DepartementSchema, Direction, Role, db
 
 
 #blueprint setup
@@ -11,10 +11,16 @@ departement = Blueprint('departement',__name__)
 def AddDepartement():
     req_Json = request.json
     name = req_Json['name']
-    idDirection = req_Json['idDirection']
-    departement = Departement(name,idDirection)
-    db.session.add(departement)
-    db.session.commit()
+    nameDirection= req_Json['nameDirection']
+    direction = Direction.query.filter_by(name=nameDirection).first()
+    idDirection = direction.id
+    departement = Departement(name,idDirection,nameDirection)
+    try:
+        db.session.add(departement)
+        db.session.commit()
+    except Exception:
+        return "0" #Name already used
+        
     return "1" #Add successfully
 
 
@@ -23,11 +29,20 @@ def AddDepartement():
 def UpdateDepartement(_id):
     req_Json = request.json
     departement = Departement.query.get(_id)
+    roles = Role.query.filter(
+    Role.nameDepartement==departement.name
+    )
     departement.name = req_Json['name']
-    departement.idDirection = req_Json['idDirection']
-    
-    db.session.commit()
-
+    departement.nameDirection = req_Json['nameDirection']
+    direction = Direction.query.filter_by(name=departement.nameDirection).first()
+    departement.idDirection = direction.id
+    try:
+        for i in roles:
+            i.nameDepartement = departement.name
+            i.idDepartement = departement.id
+        db.session.commit()
+    except Exception:
+        return "0" #Name already used
     return '1' #departement updated !!
 
 
@@ -55,7 +70,15 @@ def GetListDepartementByDirection(_id):
 @departement.route('/DeleteDepartement/<int:_id>', methods = ['DELETE'])
 def DeleteDepartement(_id):
     departement = Departement.query.get(_id)
-    db.session.delete(departement)        
-    db.session.commit()
-
-    return '1' #departement deleted !!
+    roles = Role.query.filter(
+    Role.nameDepartement==departement.name
+    )
+    x=0
+    for i in roles : 
+        x = x+1
+    if x==0:
+        db.session.delete(departement)        
+        db.session.commit()
+        return '1' #departement deleted !!
+    else :
+        return '0' #departement containes roles
